+++ src/store/actions/unverifiedTestActions.js
@@ -1,5 +1,5 @@
 import { REQUEST_UNVERIFIED_TESTS, REQUEST_UNVERIFIED_TEST, REQUEST_GRADES } from './actionTypes';
-import { getUnverifiedTests, getUnverifiedTest, getTestGrades } from '../../api/unverifiedTests-fetch';
+import { getTestsForVerification, getTestForVerification, getTestGrades } from '../../api/testsForVerification-fetch';
 
 export const setUnverifiedTests = (unverifiedTests) => ({ type: REQUEST_UNVERIFIED_TESTS, unverifiedTests });
 
@@ -8,12 +8,12 @@ export const setReports = (test) => ({ type: REQUEST_UNVERIFIED_TEST, test });
 export const setGrades = (grades) => ({ type: REQUEST_GRADES, grades });
 
 export const requestUnverifiedTests = () => (dispatch) => {
-  return getUnverifiedTests()
+  return getTestsForVerification()
     .then((data) => (dispatch(setUnverifiedTests(data))));
 };
 
 export const requestReports = (testId) => (dispatch) => {
-  return getUnverifiedTest(testId)
+  return getTestForVerification(testId)
     .then((data) => (dispatch(setReports(data))));
 };
